+++ ios/samples/hello-gltf/build-resources.sh
@@ -27,7 +27,7 @@ fi
     "${PROJECT_DIR}/../../../samples/envs/venetian_crossroads/venetian_crossroads_ibl.ktx" \
     "${PROJECT_DIR}/../../../samples/envs/venetian_crossroads/venetian_crossroads_skybox.ktx"
 
-# FilamentApp.cpp and Resources.S include files generated by resgen.
+# App.cpp and Resources.S include files generated by resgen.
 # Touch them to force Xcode to recompile.
-touch "${PROJECT_DIR}/hello-gltf/FilamentView/FilamentApp.cpp"
+touch "${PROJECT_DIR}/hello-gltf/FilamentView/App.cpp"
 touch "${PROJECT_DIR}/hello-gltf/SupportFiles/Resources.S"
